+++ far/hilight.hpp
@@ -78,7 +78,7 @@ class HighlightFiles
 
 	public:
 		void UpdateCurrentTime();
-		void GetHiColor(FileListItem **FileItem,int FileCount,bool UseAttrHighlighting=false);
+		void GetHiColor(FileListItem **FileItem,size_t FileCount,bool UseAttrHighlighting=false);
 		int  GetGroup(const FileListItem *fli);
 		void HiEdit(int MenuPos);
